+++ philo/utils/initialization.c
@@ -1,9 +1,8 @@
 #include "philo.h"
 
-uint8_t	initialize_locks_and_forks(t_table *table, t_philo *philos)
+int	initialize_locks_and_forks(t_table *table, t_philo *philos)
 {
 	int	i;
 
 	i = 0;
-
 }
